refactor(sentry): split the projection leg back under the line cap
sentry-triage-project.mjs was already past the 1,000-line hard cap in
docs/pr-checklists/recurring-review-patterns.md before this PR, and this PR
appended 99 more lines to it. scripts/ has no max-lines lint, so nothing
mechanical caught either step.

Two cuts, both along seams that needed no logic changes:

- sentry-triage-label-ensure.mjs — the settlement label self-heal. Its I/O
  is injected, so it never imports the leg that imports it, and
  runProjectionBatch keeps calling the same implementation.
- sentry-triage-project-cli.mjs — the argv surface. Help text and argument
  validation depend on nothing but the contract constants, while what stays
  behind is gh I/O and the mode drivers.

The entry module re-exports both, so the import surface is unchanged and
`node scripts/sentry-triage-project.mjs --ensure-labels …` — the invocation
the settlement fix depends on — runs exactly as before.

950 lines, from 1,226. The projection modules join the file-size cap test so
the next drift reds instead of shipping, and both new modules are routed in
the quality gate with a mapping test.
